+++ package.json
@@ -38,7 +38,6 @@
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^3.4.0",
     "firebase-tools": "^11.0.0",
-    "husky": "^7.0.0",
     "jest": "^26.6.3",
     "jest-puppeteer": "^5.0.3",
     "jsdoc": "^3.6.7",
