+++ source/DailyOverview/DailyOverview.html
@@ -2,7 +2,7 @@
 <html lang="”en”">
     <head>
         <meta charset="UTF-8" />
-        <link rel="stylesheet" href="DailyStyles.css" />
+        <link rel="stylesheet" href="./DailyStyles.css" />
         <link
             href="https://fonts.googleapis.com/css?family=Montserrat"
             rel="stylesheet"
@@ -83,8 +83,8 @@
         </div>
     </body>
 
-    <script src="./bullet.js" type="module"></script>
-    <script src="./DailyJS.js" type="module"></script>
-    <script src="./note.js" type="module"></script>
+    <script src="../General-JS/bullet.js" type="module"></script>
+    <script src="./Daily.js" type="module"></script>
+    <script src="../General-JS/note.js" type="module"></script>
     <script src="../Backend/BackendInit.js" type="module"></script>
 </html>
