--- source/Login/WeeklyOverview/Calendar.html
@@ -1,90 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Calendar</title>
-        <link rel="stylesheet" href="CalendarStyles.css" />
-        <link
-            rel="stylesheet"
-            href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css"
-        />
-    </head>
-    <body>
-        <div class="calendar-bg">
-            <div class="calendar-wrapper">
-                <div class="calendar-div"></div>
-                <a href="./WeeklyOverview.html"
-                    ><div
-                        class="fa fa-home"
-                        style="
-                            position: absolute;
-                            left: 90vw;
-                            top: 2vh;
-                            font-size: 5vmax;
-                            color: #585858;
-                        "
-                    ></div
-                ></a>
-                <div id="goalReminder">
-                    <div>
-                        <div id="goalContainer">
-                            <h2>Monthly Goals:</h2>
-                            <h2 id="plus-month">+</h2>
-                        </div>
-                        <div id="monthGoal">
-                            <ul>
-                                <li>Task 1</li>
-                                <li>Task 2</li>
-                            </ul>
-                            <form class="entry-form-hide" id="month-form">
-                                <input
-                                    type="text"
-                                    placeholder="New Bullet Text"
-                                    class="entry-form-text"
-                                    id="entry"
-                                />
-                                <input
-                                    type="submit"
-                                    value="Post"
-                                    class="entry-form-button"
-                                    id="entry-button"
-                                />
-                            </form>
-                        </div>
-                    </div>
-                    <div>
-                        <div id="goalContainer">
-                            <h2>Yearly Goals:</h2>
-                            <h2 id="plus-year">+</h2>
-                        </div>
-                        <div id="yearGoal">
-                            <form class="entry-form-hide" id="year-form">
-                                <input
-                                    type="text"
-                                    placeholder="New Bullet Text"
-                                    class="entry-form-text"
-                                    id="entry"
-                                />
-                                <input
-                                    type="submit"
-                                    value="Post"
-                                    class="entry-form-button"
-                                    id="entry-button"
-                                />
-                            </form>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div id="calendar-popup">
-            <h3 id="popup-title">Enter New Goal</h3>
-            <div id="close-popup">&times;</div>
-            <input type="text" class="popup-text" id="calendar-popup-text"/>
-            <button id="submit-note">OK</button>
-            <button id="cancel-note">Cancel</button>
-        </div>
-    </body>
-
-    <script src="WeeklyJS.js" type="module"></script>
-    <script src="Calendar.js" type="module"></script>
-</html>
