+++ source/WeeklyOverview/WeeklyOverview.html
@@ -2,7 +2,7 @@
 <html>
     <head>
         <title>Weekly Overview</title>
-        <link rel="stylesheet" href="WeeklyStyles.css" />
+        <link rel="stylesheet" href="./WeeklyStyles.css" />
         <link
             href="https://fonts.googleapis.com/css?family=Montserrat"
             rel="stylesheet"
@@ -88,7 +88,7 @@
                         <form action="#" method="#">
                             <div class="passField">
                                 <img
-                                    src="../../Images/password-icon.png"
+                                    src="../Images/password-icon.png"
                                     alt=""
                                     class="pwd-icon"
                                 />
@@ -100,7 +100,7 @@
                                     value=""
                                 />
                                 <img
-                                    src="../../Images/password-icon.png"
+                                    src="../Images/password-icon.png"
                                     alt=""
                                     class="pwd-icon"
                                 />
@@ -112,7 +112,7 @@
                                     value=""
                                 />
                                 <img
-                                    src="../../Images/password-icon.png"
+                                    src="../Images/password-icon.png"
                                     alt=""
                                     class="pwd-icon"
                                 />
@@ -163,9 +163,9 @@
                 </div>
             </div>
         </div>
-        <script src="../../Backend/BackendInit.js" type="module"></script>
-        <script src="../../Backend/FirebaseInit.js" type="module"></script>
-        <script src="../WeeklyOverview/note.js" type="module"></script>
-        <script src="../WeeklyOverview/WeeklyJS.js" type="module"></script>
+        <script src="../Backend/BackendInit.js" type="module"></script>
+        <script src="../Backend/FirebaseInit.js" type="module"></script>
+        <script src="../General-JS/note.js" type="module"></script>
+        <script src="./Weekly.js" type="module"></script>
     </body>
 </html>
