+++ source/WeeklyOverview/WeeklyStyles.css
@@ -38,18 +38,19 @@ body {
     padding: 0.5%;
     color: white;
     font-size: 7vh;
-    margin-left: 7%;
-    margin-top: 2%;
-    padding: 5px;
+    /* margin-left: 7%;
+    margin-top: 2%; */
+    /* padding: 5px; */
     left: 7%;
     top: 7%;
     background-color: rgba(30, 30, 30, 0.7);
     border-radius: 10px;
 }
 #header_date:hover {
-    cursor: pointer;
+    /* border: 1px solid #888; */
     background-color: #ddd;
     color: #888;
+    cursor: pointer;
 }
 
 #header_button_container {
